fix(mla): cast advisory lock key so postgres resolves the int4 overload

+++ src/lib/mlaPersistence.ts
@@ -5,8 +5,12 @@ export const lockMlaAssignment = async (
   chainId: number,
   market: string,
 ) => {
+  // Prisma binds JS numbers as int8, but the two-key advisory lock overload
+  // is (int4, int4) and Postgres does not downcast during function
+  // resolution - without the explicit cast the call fails with 42883.
+  // hashtext() already returns int4.
   await transaction.$queryRaw`
-    SELECT pg_advisory_xact_lock(${chainId}, hashtext(${market}))
+    SELECT pg_advisory_xact_lock((${chainId})::int, hashtext(${market}))
   `
 }
